+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { SVGLogo } from "./image";
 export { MajangText } from "./text";
-export { BottomButton } from "./bottomButton";
+export * from "./common"
 export { ClickableText } from "./clickableText";
